Fix a lot of checkstyle warnings.
Some of the remaining warnings should be removed from the checkstyle rule set.
