+++ src/main/java/datastructure/Node.java
@@ -5,7 +5,7 @@ import javafx.scene.shape.Rectangle;
 /**
  * Created by 101010.
  */
-public class Node extends Rectangle{
+public class Node extends Rectangle {
     /**
      * The length of the segment corresponding to the node.
      */
@@ -107,7 +107,7 @@ public class Node extends Rectangle{
     }
 
     /**
-     *
+     * Equals method which return true if two nodes are exactly the same.
      * @param other the object to compare to.
      * @return whether this Node is equal to the other object.
      */
