+++ src/main/java/datastructure/SegmentDB.java
@@ -8,8 +8,6 @@ import java.sql.DriverManager;
 import java.sql.Connection;
 import java.sql.ResultSet;
 import java.sql.SQLException;
-import java.util.LinkedList;
-import java.util.Queue;
 
 /**
  * Created by 101010.
