+++ src/main/java/window/App.java
@@ -3,15 +3,10 @@ package window;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Parent;
 import javafx.scene.Scene;
-import javafx.scene.text.Text;
-import javafx.scene.layout.GridPane;
 import javafx.stage.Stage;
 
 import java.io.IOException;
-import javafx.stage.FileChooser;
-import javafx.scene.control.Button;
 
-import java.io.File;
 /**
  * Created by 101010 on 8-5-2017.
  */
@@ -19,7 +14,7 @@ public class App extends javafx.application.Application {
 
     /**
      * Method that starts the application.
-     * @param args
+     * @param args Parameters to run this application with.
      */
     public static void main(final String[] args) {
         launch(args);
